Added template argument for SSLClient class.
